Extracted Youtube and YT MP3 download methods to separate YoutubeService
