+++ BeatSage-Downloader-WPF/Models/DownloadManager.cs
@@ -13,18 +13,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using MahApps.Metro.Controls;
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
-using System.Diagnostics;
-using System.Runtime.Serialization.Formatters.Binary;
 using YoutubeExplode;
 using YoutubeExplode.Videos.Streams;
 
@@ -33,20 +21,16 @@ namespace BeatSage_Downloader
     [Serializable]
     public class DownloadManager
     {
+        //Fields
         public static ObservableCollection<Download> downloads = new ObservableCollection<Download>();
-
         public static readonly HttpClient httpClient = new HttpClient();
-
-        private static DataGrid dataGrid;
-
         public static CancellationTokenSource cts;
-
         public static Label newUpdateAvailableLabel;
-
-        public DownloadManager(DataGrid newDataGrid)
+        public static ObservableCollection<Download> Downloads { get; }
+        
+        //Constructor
+        public DownloadManager()
         {
-            dataGrid = newDataGrid;
-
             httpClient.DefaultRequestHeaders.Add("Host", "beatsage.com");
             httpClient.DefaultRequestHeaders.Add("Accept", "*/*");
             httpClient.DefaultRequestHeaders.Add("User-Agent", "BeatSage-Downloader/1.2.1");
@@ -65,11 +49,10 @@ namespace BeatSage_Downloader
 
         }
 
+        //Methods
         public async void RunDownloads()
         {
             Console.WriteLine("RunDownloads Started");
-
-            int previousNumberOfDownloads = downloads.Count;
 
             while (true)
             {
@@ -102,11 +85,11 @@ namespace BeatSage_Downloader
                         {
                             if (Properties.Settings.Default.enableLocalYouTubeDownload)
                             {
-                                await CreateCustomLevelWithLocalMP3Download(itemUrl, currentDownload);
+                                await YoutubeService.CreateCustomLevelWithLocalMP3Download(itemUrl, currentDownload, httpClient, cts);
                             }
                             else
                             {
-                                await RetrieveMetaData(itemUrl, currentDownload);
+                                await DownloadManager.RetrieveMetaData(itemUrl, currentDownload);
                             }
                         }
                         catch
@@ -123,7 +106,7 @@ namespace BeatSage_Downloader
                     {
                         try
                         {
-                            await CreateCustomLevelFromFile(currentDownload);
+                            await DownloadManager.CreateCustomLevelFromFile(currentDownload);
                         }
                         catch
                         {
@@ -138,9 +121,8 @@ namespace BeatSage_Downloader
 
                 cts.Dispose();
                 System.Threading.Thread.Sleep(1000);
+
             }
-
-
         }
 
         public static async Task CheckUpdateAvailable()
@@ -184,134 +166,14 @@ namespace BeatSage_Downloader
             }
         }
 
-
-        public static ObservableCollection<Download> Downloads
-        {
-            get
-            {
-                return downloads;
-            }
-        }
-
         public void Add(Download download)
         {
             downloads.Add(download);
         }
 
-        public async static Task CreateCustomLevelWithLocalMP3Download(string url, Download download)
+        public static async Task CreateCustomLevelWithLocalMP3Download(string url, Download download)
         {
-            download.Status = "Downloading File";
-
-            var youtube = new YoutubeClient();
-
-            // You can specify video ID or URL
-            var video = await youtube.Videos.GetAsync(url);
-
-            var duration = video.Duration; // 00:07:14
-
-            if (video.Duration.Minutes + (video.Duration.Seconds/60) > 10)
-            {
-                return;
-            }
-
-            string artistName = "Unknown";
-            string trackName = "Unknown";
-
-            var invalids = System.IO.Path.GetInvalidFileNameChars();
-
-            if (video.Author != null)
-            {
-                artistName = String.Join("_", video.Author.Split(invalids, StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
-            }
-
-            if (video.Title != null)
-            {
-                trackName = String.Join("_", video.Title.Split(invalids, StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
-            }
-
-            download.Artist = artistName;
-            download.Title = trackName;
-
-            string fileName = "[BSD] " + trackName + " - " + artistName;
-            download.FilePath = fileName + ".mp3";
-
-            if (!Properties.Settings.Default.overwriteExisting)
-            {
-                if (((!Properties.Settings.Default.automaticExtraction) && (File.Exists(Properties.Settings.Default.outputDirectory + @"\" + fileName + ".zip"))) || ((Properties.Settings.Default.automaticExtraction) && (Directory.Exists(Properties.Settings.Default.outputDirectory + @"\" + fileName))))
-                {
-                    download.Status = "Already Exists";
-                    download.IsAlive = false;
-                    return;
-                }
-            }
-            
-            var streamManifest = await youtube.Videos.Streams.GetManifestAsync(video.Id);
-
-            var streamInfo = streamManifest.GetAudioOnly().WithHighestBitrate();
-
-            if (streamInfo != null)
-            {
-                // Get the actual stream
-                var stream = await youtube.Videos.Streams.GetAsync(streamInfo);
-
-                // Download the stream to file
-                await youtube.Videos.Streams.DownloadAsync(streamInfo, fileName + ".mp3");
-            }
-
-            string boundary = "----WebKitFormBoundaryaA38RFcmCeKFPOms";
-            var content = new MultipartFormDataContent(boundary);
-
-            byte[] bytes = System.IO.File.ReadAllBytes(download.FilePath);
-
-            if (File.Exists(download.FilePath))
-            {
-                File.Delete(download.FilePath);
-            }
-            
-            content.Add(new ByteArrayContent(bytes), "audio_file", download.FilePath);
-
-            using (WebClient client = new WebClient())
-            {
-                client.DownloadFile(new Uri("https://img.youtube.com/vi/" + video.Id + "/maxresdefault.jpg"), "cover.jpg");
-            }
-
-            byte[] imageData = System.IO.File.ReadAllBytes("cover.jpg");
-
-            if (imageData != null)
-            {
-                var imageContent = new ByteArrayContent(imageData);
-                imageContent.Headers.Remove("Content-Type");
-                imageContent.Headers.Add("Content-Disposition", "form-data; name=\"cover_art\"; filename=\"cover\"");
-                imageContent.Headers.Add("Content-Type", "image/jpeg");
-                content.Add(imageContent);
-            }
-
-            if (File.Exists("cover.jpg"))
-            {
-                File.Delete("cover.jpg");
-            }
-
-            content.Add(new StringContent(trackName), "audio_metadata_title");
-            content.Add(new StringContent(artistName), "audio_metadata_artist");
-            content.Add(new StringContent(download.Difficulties), "difficulties");
-            content.Add(new StringContent(download.GameModes), "modes");
-            content.Add(new StringContent(download.SongEvents), "events");
-            content.Add(new StringContent(download.Environment), "environment");
-            content.Add(new StringContent(download.ModelVersion), "system_tag");
-
-            var response = await httpClient.PostAsync("https://beatsage.com/beatsaber_custom_level_create", content, cts.Token);
-
-            var responseString = await response.Content.ReadAsStringAsync();
-
-            Console.WriteLine(responseString);
-
-            JObject jsonString = JObject.Parse(responseString);
-
-            string levelID = (string)jsonString["id"];
-
-            Console.WriteLine(levelID);
-
-            await CheckDownload(levelID, trackName, artistName, download);
+            await YoutubeService.CreateCustomLevelWithLocalMP3Download(url, download, httpClient, cts);
         }
 
         public async static Task RetrieveMetaData(string url, Download download)
@@ -458,14 +320,13 @@ namespace BeatSage_Downloader
             await CheckDownload(levelID, trackName, artistName, download);
         }
 
-        static async Task CreateCustomLevelFromFile(Download download)
+        public static async Task CreateCustomLevelFromFile(Download download)
         {
             download.Status = "Uploading File";
 
             TagLib.File tagFile = TagLib.File.Create(download.FilePath);
 
             string artistName = "Unknown";
-            string trackName = "Unknown";
             byte[] imageData = null;
 
             var invalids = System.IO.Path.GetInvalidFileNameChars();
@@ -475,6 +336,7 @@ namespace BeatSage_Downloader
                 artistName = String.Join("_", tagFile.Tag.FirstPerformer.Split(invalids, StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
             }
 
+            string trackName;
             if (tagFile.Tag.Title != null)
             {
                 trackName = String.Join("_", tagFile.Tag.Title.Split(invalids, StringSplitOptions.RemoveEmptyEntries)).TrimEnd('.');
@@ -547,7 +409,7 @@ namespace BeatSage_Downloader
             await CheckDownload(levelID, trackName, artistName, download);
         }
 
-        static async Task CheckDownload(string levelId, string trackName, string artistName, Download download)
+        public static async Task CheckDownload(string levelId, string trackName, string artistName, Download download)
         {
             download.Status = "Generating Custom Level";
 
@@ -580,12 +442,10 @@ namespace BeatSage_Downloader
                     JObject jsonString = JObject.Parse(responseString);
 
                     levelStatus = (string)jsonString["status"];
-
                 }
                 catch
                 {
                 }
-
             }
 
             if (levelStatus == "DONE")
@@ -649,66 +509,6 @@ namespace BeatSage_Downloader
 
             download.Status = "Completed";
             download.IsAlive = false;
-        }
-
-        public static List<string> RetrieveYouTubePlaylist(string playlistULR)
-        {
-            string cleanPlaylistURL = playlistULR.Replace("watch?v", "playlist?v").Replace("music.", "");
-
-            string htmlContent = new WebClient().DownloadString(cleanPlaylistURL);
-
-            List<string> urls = new List<string>();
-
-            string searchString = "data-video-id=";
-
-            int htmlPointerLocation = 1;
-
-            while (htmlPointerLocation > 0)
-            {
-
-                htmlPointerLocation = htmlContent.IndexOf(searchString);
-
-                if (htmlPointerLocation > 0)
-                {
-                    string temporaryURL = "";
-
-                    int i;
-                    for (i = (htmlPointerLocation + searchString.Count() + 1); i < htmlContent.Count(); i++)
-                    {
-                        if (htmlContent[i].ToString() != "\"")
-                        {
-                            temporaryURL += htmlContent[i];
-                        }
-                        else
-                        {
-                            break;
-                        }
-                    }
-
-                    htmlContent = htmlContent.Substring(i);
-
-                    string newURL = "https://www.youtube.com/watch?v=" + temporaryURL;
-
-                    bool alreadyExists = false;
-
-                    foreach (string currentURL in urls)
-                    {
-                        if ((currentURL == newURL) || (newURL.Contains(currentURL)))
-                        {
-                            alreadyExists = true;
-                            break;
-                        }
-                    }
-
-                    if (alreadyExists == false)
-                    {
-                        Console.WriteLine(newURL);
-                        urls.Add(newURL);
-                    }
-                }
-            }
-
-            return urls;
         }
     }
 }
